rename types/relayer to types/executor & feat: testingMode
