+++ packages/cli/src/cmds/start/index.ts
@@ -5,14 +5,21 @@ import { bundlerHandler } from "./handler";
 
 export interface IBundlerArgs {
   networks: NetworkName;
+  testingMode: boolean;
 }
 
 export const bundlerOptions = {
   networks: {
     description: "Name of the EVM chain to join",
     type: "string[]",
-    defaultDescription: "goerli",
+    default: "goerli",
     choices: networkNames,
+  },
+  testingMode: {
+    description: "Run bundler in dev mode (For testing against test suite)",
+    type: "boolean",
+    default: false,
+    choices: [true, false],
   },
 };
 
@@ -21,7 +28,7 @@ export const start: ICliCommand<IBundlerArgs, IGlobalArgs> = {
   describe: "Run a bundler client",
   examples: [
     {
-      command: "start --network goerli",
+      command: "start",
       description: "Run a bundler client and connect to the goerli testnet",
     },
   ],
