+++ packages/executor/src/modules/eth.ts
@@ -6,13 +6,13 @@ import {
   EntryPoint,
   UserOperationEventEvent,
   UserOperationStruct,
-} from "types/lib/relayer/contracts/EntryPoint";
+} from "types/lib/executor/contracts/EntryPoint";
 import {
   EstimatedUserOperationGas,
   UserOperationByHashResponse,
   UserOperationReceipt,
 } from "types/lib/api/interfaces";
-import { EntryPoint__factory } from "types/lib/relayer/contracts/factories";
+import { EntryPoint__factory } from "types/lib/executor/contracts/factories";
 import { NetworkConfig } from "../config";
 import { deepHexlify, packUserOp } from "../utils";
 import { UserOpValidationService, MempoolService } from "../services";
